Refactor: move global config class into its own file
Part of the ongoing breakup of binding.py
