+++ opsramp/binding.py
@@ -23,6 +23,7 @@ from __future__ import print_function
 import base64
 
 from opsramp.base import ApiObject, ApiWrapper
+from opsramp.globalconfig import GlobalConfig
 
 
 def connect(url, key, secret):
@@ -57,41 +58,6 @@ class Opsramp(ApiWrapper):
 
     def tenant(self, name):
         return Tenant(self, name)
-
-
-class GlobalConfig(ApiWrapper):
-    def __init__(self, parent):
-        super(GlobalConfig, self).__init__(parent.api, '')
-
-    def get_alert_types(self):
-        return self.api.get('/alertTypes')
-
-    def get_countries(self):
-        return self.api.get('/cfg/countries')
-
-    def get_timezones(self):
-        return self.api.get('/cfg/timezones')
-
-    def get_alert_technologies(self):
-        return self.api.get('/cfg/alertTechnologies')
-
-    def get_channels(self):
-        return self.api.get('/cfg/tenants/channels')
-
-    def get_nocs(self):
-        # Bizarrely this API call throws a 500 error if there are
-        # no NOCs defined. Handle it gracefully.
-        try:
-            retval = self.api.get('/cfg/tenants/nocs')
-        except RuntimeError as e:
-            if '"code":"0005"' in str(e):
-                retval = []
-            else:
-                raise
-        return retval
-
-    def get_device_types(self):
-        return self.api.get('/cfg/devices/types')
 
 
 class Tenant(ApiWrapper):
